Created Length system
After all the letters have been assigned numbers, each number will get a 'length number', where if the number is ten or above, it will get a 2. Otherwise, it will get a 1. Then, every other index of new_sentence [0, 2, 4, 6, etc], it will put one of those numbers (in order). Then, at the very end of the final sequence. The length of the sequence will be put. So that the code knows how many numbers were in the original sequence