Added animation config setting
Added the ability for, when you put True in the config under animation, it'll show the information. If you put False, then it won't animate the text. If something is entered that the program doesn't understand (anything that isn't True or False), it'll show an error message and then exit the program.