Fixed issue with date-time stamping
I fixed an issue with date-time stamping, in which days of dates which were one digit long (0-9) would crash the program, while 2 digit dates (10-31) would be fine. To fix this, I just added an if-else statement so that if the length was more than 2 digits, it would do what I had made before, else, it would use the int of the date as the key for the dictionary.